requirements.txt
paho-mqtt>=1.6.0,<2.0.0
flask>=2.0.0
waitress>=2.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
